Add more parameters to promo modal html
